Updated something idk

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: hbreeze <hbreeze@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/27 20:40:30 by hbreeze           #+#    #+#             */
-/*   Updated: 2024/08/31 15:35:17 by hbreeze          ###   ########.fr       */
+/*   Updated: 2025/03/07 18:05:18 by hbreeze          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ static int	is_whitespace(char c)
 
 int	ft_atoi(const char *nptr)
 {
-	int	output;
+	long	output;
 	int	sign;
 
 	output = 0;
@@ -40,5 +40,5 @@ int	ft_atoi(const char *nptr)
 		output += *nptr - '0';
 		nptr++;
 	}
-	return (output * sign);
+	return ((int)(output * sign));
 }
